Genie Home Assignment/pyproject.toml
[tool.poetry]
name = "home_assignment"
version = "0.0.1"
description = "Home assignment"
authors = [

]
maintainers = [

]

readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.78.0"
uvicorn = { version = "^0.17.6", extras = ["standard"] }
SQLAlchemy = {extras = ["mypy"], version = "^1.4.40"}
asyncpg = {version = "^0.26.0", extras = ["sa"]}
aiofiles = "^0.8.0"
httptools = "^0.4.0"
greenlet = "^1.1.2"
strawberry-graphql = {extras = ["fastapi"], version = "^0.138.1"}
httpx = "^0.23.0"
python-jose = "^3.3.0"
yappi = "^1.3.6"
arrow = "^1.2.3"
psycopg2-binary = "^2.9.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
